requirements.txt
# Core backend
fastapi
uvicorn
requests
python-dotenv
pydantic

# LangChain + OpenAI
langchain
langchain-core
langchain-openai


